Added a method for checking for changes
I've added a method for checking for changes in sublists when doing changes to a sublist of the internal list. This method will then determine if the list has been modified, increments the modification count accordingly, and fires ListDataEvents that cover the ranges of altered elements. This way, if a sublist is altered, but only part of it is changed, the listeners will only be notified of the part that was altered instead of the whole sublist.