Updated ranged sort method
I've updated the ranged version of the sort method for ArrayListModel. The method now creates a copy of the sublist to be sorted before sorting it, so that the sorted sublist can be compared with the unmodified version to detect any changes. After sorting the sublist, it will now only fire contents changed events on the parts of the sublist that have been altered by sorting the sublist. I've also added documentation to the method.