Update ArrayListModel.java
I added a throw clause into the documentation for the ranged sort method stating that an UnsupportedOperationException would be thrown if the list is unmodifiable. This may be unnecessary, though, since the ranged sort method works on a sublist of the internal ArrayList, and ArrayLists are modifiable. However, this method could be overridden in the process of making an unmodifiable ArrayListModel.