Update ArrayListModel.java
I've updated the documentation for the replaceRange and ranged sort method. I've also added checks to the replaceRange, ranged sort, removeRange, and private batchRemove methods for whether the list is empty, and return false if it is since they can't do anything if the list is empty. I've also moved the check for whether both the filter and collection are null in the private batchRemove method to before the check if the indexes are the same or the list is empty, since the null pointer exception should probably be fired first.