Added documentation to the private batchRemove
I've added documentation to the private batchRemove method. This method is used for removing elements depending on whether they match a given filter or are in a given collection, since the removeIf, removeAll, and retainAll methods are all similar, with the only real difference being how they determine whether an element is to be removed.