Fixed removeIf dealing with the selected item
There was a bug where if the model still contained the selected item after the items were removed, then it would clear the selected item. The exact opposite behavior was intended (the selected item should be cleared if the selected item is no longer in the list). This has been fixed.